Work around broken MIPS build
We used to hardcode arm/util/dwarf-regs.c for all platforms but it
doesn't look like there is one for MIPS yet. Add a fake one for MIPS for
now and when used it will raise an error at runtime.

Also remove an extra util/config.c from the makefile.

Change-Id: I035110be302050d72a6ba0bee6bd2979cb5d0056