[PATCH] perf tool: fix bug in usage of the basename() function
The basename() implementation varies a lot between systems.
The Linux man page says: "basename may modify the content of the path,
so it may be desirable to pass a copy when calling the function".

The dso__set_basename() function was not doing this causing problems
on some systems with wrong library names being shown by perf report.

This patch fixes the problem.

Change-Id: I1501adff2c092f7863f8673892bea93f2cb74563
Reported-by: Ben Cheng <bccheng@google.com>
Signed-off-by: Stephane Eranian <eranian@google.com>