Fix perf(1).
This is still a mess, but it's less of a mess and at least the perf basics
now work. I've also added a couple of TODOs (and raised an internal bug)
for one problem, and we have an existing internal bug for the strerror_r
problem, and I've raised a new internal bug for the fact that 'perfhost
report' doesn't seem to be able to read our symbols.

(cherry-pick of 708e65a9521667eeba62945e418deabc70423f52.)

Bug: 16705027
Change-Id: I6eb321162b48f0625901e5be5d7915db7294b8e3