Sync with upstream perf-3.12.
Modulo a handful of changes (now marked with #if defined(ANDROID_PATCHES)),
this brings us in line with upstream kernel source. I'll look at adding
a GNU-style basename(3) to bionic and seeing if we can fix libelf to not
need <stdio_ext.h> (or add <stdio_ext.h>) and then come back to remove
more of the hacks.

I also got rid of a lot of the stuff that we can just get from our uapi
headers, rather than have a second copy here.

Change-Id: Ifcddea46708471c86ee6fee6f31295875c7308bb