Remove the perfhost target.
The perfhost target isn't useful, because you already have perf installed.
What's more, we were never building the whole thing anyway and were dlopen'ing
parts of the one you already had installed. Let's stop kidding ourselves and
just admit that we're happy with the one on the host. (IDEs are going to need
something that works on Mac OS and Windows anyway, so this isn't useful for
them.)

Change-Id: I8018dfc5645ec5d01c257af6f461e1e3db30fb53